fix: move openBrowser from SDK to code package to prevent double browser open
startLocalAuthServer called both onAuthUrl callback and openBrowser,
causing the browser to open twice (e.g. in VS Code Remote SSH where
onAuthUrl uses vscode.env.openExternal and SDK uses xdg-open).

Moved openBrowser to LoginCommand.tsx so the CLI frontend controls
browser opening. SDK now only notifies via onAuthUrl callback.

Fixes #1267
